+++ src/components/study/EmissionSourceContributorForm.tsx
@@ -9,10 +9,9 @@ import { getEmissionFactorValue } from '@/utils/emissionFactors'
 import { formatEmissionFactorNumber } from '@/utils/number'
 import AddIcon from '@mui/icons-material/Add'
 import { TextField } from '@mui/material'
-import { Import, StudyResultUnit, SubPost, Unit } from '@prisma/client'
+import { StudyResultUnit, SubPost, Unit } from '@prisma/client'
 import classNames from 'classnames'
 import { useTranslations } from 'next-intl'
-import { useMemo } from 'react'
 import { Path } from 'react-hook-form'
 import LinkButton from '../base/LinkButton'
 import QualitySelect from '../form/QualitySelect'
@@ -25,7 +24,8 @@ interface Props {
   selectedFactor?: EmissionFactorWithMetaData
   subPost: SubPost
   update: (key: Path<UpdateEmissionSourceCommand>, value: string | number | boolean | null) => void
-  studyImportVersions: { id: string; source: Import; importVersionId: string }[]
+  isFromOldImport: boolean
+  currentBCVersion: string
 }
 
 const getDetail = (metadata: Exclude<EmissionFactorWithMetaData['metaData'], undefined>) =>
@@ -37,30 +37,12 @@ const EmissionSourceContributorForm = ({
   subPost,
   selectedFactor,
   update,
-  studyImportVersions,
+  isFromOldImport,
+  currentBCVersion,
 }: Props) => {
   const t = useTranslations('emissionSource')
   const tResultUnits = useTranslations('study.results.units')
   const tUnits = useTranslations('units')
-
-  const isFromOldImport = useMemo(
-    () =>
-      !!selectedFactor?.version?.id &&
-      !studyImportVersions
-        .map((studyImportVersion) => studyImportVersion.importVersionId)
-        .includes(selectedFactor.version.id),
-    [selectedFactor, studyImportVersions],
-  )
-
-  const currentBCVersion = useMemo(() => {
-    const versionId = isFromOldImport
-      ? studyImportVersions.find((studyImportVersion) => studyImportVersion.source === Import.BaseEmpreinte)
-          ?.importVersionId || ''
-      : ''
-    return isFromOldImport
-      ? emissionFactors.find((factor) => factor?.version?.id === versionId)?.version?.name || ''
-      : ''
-  }, [studyImportVersions, isFromOldImport, emissionFactors])
 
   return (
     <>
