+++ src/components/study/EmissionSourceForm.tsx
@@ -23,7 +23,6 @@ import { FormControl, InputLabel, MenuItem, TextField } from '@mui/material'
 import {
   EmissionSourceCaracterisation,
   EmissionSourceType,
-  Import,
   StudyResultUnit,
   StudyRole,
   SubPost,
@@ -65,7 +64,8 @@ interface Props {
   mandatoryCaracterisation: boolean
   status: EmissionSourcesStatus
   studySites: FullStudy['sites']
-  studyImportVersions: { id: string; source: Import; importVersionId: string }[]
+  isFromOldImport: boolean
+  currentBCVersion: string
 }
 
 const EmissionSourceForm = ({
@@ -83,7 +83,8 @@ const EmissionSourceForm = ({
   mandatoryCaracterisation,
   status,
   studySites,
-  studyImportVersions,
+  isFromOldImport,
+  currentBCVersion,
 }: Props) => {
   const t = useTranslations('emissionSource')
   const tUnits = useTranslations('units')
@@ -114,25 +115,6 @@ const EmissionSourceForm = ({
   const specificFEDefaultQuality = specificFEQualities.find((quality) => quality)
   const canShrinkSpecificFEQuality =
     !specificFEDefaultQuality || specificFEQualities.every((quality) => quality === specificFEDefaultQuality)
-
-  const isFromOldImport = useMemo(
-    () =>
-      !!selectedFactor?.version?.id &&
-      !studyImportVersions
-        .map((studyImportVersion) => studyImportVersion.importVersionId)
-        .includes(selectedFactor.version.id),
-    [selectedFactor, studyImportVersions],
-  )
-
-  const currentBCVersion = useMemo(() => {
-    const versionId = isFromOldImport
-      ? studyImportVersions.find((studyImportVersion) => studyImportVersion.source === Import.BaseEmpreinte)
-          ?.importVersionId || ''
-      : ''
-    return isFromOldImport
-      ? emissionFactors.find((factor) => factor?.version?.id === versionId)?.version?.name || ''
-      : ''
-  }, [studyImportVersions, isFromOldImport, emissionFactors])
 
   const handleUpdate = (event: React.FocusEvent<HTMLInputElement | HTMLTextAreaElement>) => {
     if (Number(event.target.value) > 0) {
